Some restructuring of utilities
