+++ apps/extension/src/background/chains/service.ts
@@ -1,5 +1,6 @@
-import { KVStore, Debouncer } from "@anoma/storage";
 import { ChainInfo as Chain } from "@keplr-wallet/types";
+import { KVStore } from "@anoma/storage";
+import { debounce } from "@anoma/utils";
 import { Env } from "../../router";
 
 type ChainRemovedHandler = (chainId: string, identifier: string) => void;
@@ -19,7 +20,7 @@ export class ChainsService {
     console.info("ChainsService initialized");
   }
 
-  readonly getChains: () => Promise<any[]> = Debouncer.promise(async () => {
+  readonly getChains: () => Promise<any[]> = debounce(async () => {
     if (this.cachedChains) {
       return this.cachedChains;
     }
