+++ apps/namada-interface/src/utils/encoding/index.ts
