+++ packages/crypto/lib/Cargo.toml
@@ -2,6 +2,9 @@
 name = "crypto"
 version = "0.1.0"
 edition = "2021"
+repository = "https://github.com/anoma/namada-interface/"
+description = "Cryptographic functions for Anoma apps"
+license = "MIT"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
