+++ packages/storage/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./store";
-export * from "./utils";
