+++ packages/types/src/crypto.ts
@@ -1,0 +1,16 @@
+export type RNG = <
+  T extends
+    | Int8Array
+    | Int16Array
+    | Int32Array
+    | Uint8Array
+    | Uint16Array
+    | Uint32Array
+    | Uint8ClampedArray
+    | Float32Array
+    | Float64Array
+    | DataView
+    | null
+>(
+  array: T
+) => Promise<T>;
