+++ packages/types/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./anoma";
+export * from "./crypto";
