+++ packages/utils/package.json
@@ -12,11 +12,11 @@
     "lint:fix": "yarn lint -- --fix"
   },
   "dependencies": {
-    "@anoma/masp-web": "0.1.0",
     "@cosmjs/json-rpc": "^0.27.1",
     "bs58": "^5.0.0",
     "crypto-js": "^4.1.1",
-    "luxon": "^3.0.1"
+    "luxon": "^3.0.1",
+    "sha.js": "^2.4.11"
   },
   "devDependencies": {
     "@types/bs58": "^4.0.1",
@@ -24,6 +24,7 @@
     "@types/jest": "^28.1.8",
     "@types/luxon": "^3.0.0",
     "@types/node": "^18.7.13",
+    "@types/sha.js": "^2.4.0",
     "eslint": "^8.22.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.5.0",
